fix: 변경된 파일 컨벤션 적용
