+++ src/App.tsx
@@ -2,7 +2,6 @@ import './App.css'
 import { Routes, Route } from 'react-router-dom'
 import Login from './pages/app/Login'
 import Main from './pages/app/Main'
-import MyPage from './pages/app/mypage/MyPage'
 import { Fragment, useEffect, useState } from 'react'
 import SplashScreen from './components/SplashScreen'
 import { cn } from './lib/utils'
@@ -18,11 +17,12 @@ import ManageTeam from './pages/app/challenges/[id]/teams/[id]/joined'
 import TeamModify from './pages/app/challenges/[id]/teams/[id]/modify'
 import ChallengeSubmitTeam from './pages/app/challenges/[id]/submit/team/[teamId]'
 import { Toaster } from './components/ui/sonner'
-import MyPoints from '@/pages/app/mypage/MyPoints'
-import WithDraw from '@/pages/app/mypage/WithDraw'
-import EditProfile from '@/pages/app/mypage/EditProfile'
-import CertifiedChallenges from '@/pages/app/mypage/CertifiedChallenges'
-import CertifiedChallengesDetail from '@/pages/app/mypage/CertifiedChallengesDetail'
+import MyPoints from '@/pages/app/my-page/my-points'
+import WithDraw from '@/pages/app/my-page/withdraw'
+import EditProfile from '@/pages/app/my-page/edit-profile'
+import CertifiedChallenges from '@/pages/app/my-page/certifed-challenges'
+import CertifiedChallengesDetail from '@/pages/app/my-page/certified-challenges-detail'
+import MyPage from '@/pages/app/my-page/my-page'
 
 const queryClient = new QueryClient()
