+++ src/components/challenge-type-switch.tsx
