+++ src/components/my-page-screen/logo-and-name.tsx
