+++ src/components/my-page-screen/points-history-item.tsx
