+++ src/components/withdraw-screen/notice-message.tsx
