+++ tsconfig.app.json
@@ -31,5 +31,5 @@
     "noUnusedParameters": true,
     "esModuleInterop": true
   },
-  "include": ["src"]
+  "include": ["src", "src/components/withdraw-screen/.tsx"]
 }
